requirements.txt
yfinance>=0.2.18
pandas>=2.0.0
requests>=2.31.0
fredapi>=0.5.0
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.25.0
numpy>=1.24.0
lxml>=4.9.0
langgraph>=0.2.0
openai>=1.0.0
scipy>=1.11.0
ta>=0.10.0
python-dotenv>=1.0.0 
sqlalchemy>=2.0.0
sqlite-utils>=3.36.0 
matplotlib>=3.7.0
seaborn>=0.12.0